pyproject.toml
[project]
name = "midi-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Basti",email = "a@b.de"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "pygame (>=2.6.1,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "tkinterdnd2 (>=0.4.3,<0.5.0)",
    "ruff (>=0.14.14,<0.15.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "pyinstaller (>=6.18.0,<7.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
